+++ packages/@expo/config/src/paths/paths.ts
@@ -2,7 +2,9 @@ import fs from 'fs';
 import path from 'path';
 import resolveFrom from 'resolve-from';
 
+import { env } from './env';
 import { getBareExtensions } from './extensions';
+import { findWorkspaceRoot } from './workspaces';
 import { getPackageJson } from '../Config';
 import { PackageJSONConfig } from '../Config.types';
 import { ConfigError } from '../Errors';
@@ -11,7 +13,7 @@ import { ConfigError } from '../Errors';
 export function ensureSlash(inputPath: string, needsSlash: boolean): string {
   const hasSlash = inputPath.endsWith('/');
   if (hasSlash && !needsSlash) {
-    return inputPath.substr(0, inputPath.length - 1);
+    return inputPath.substring(0, inputPath.length - 1);
   } else if (!hasSlash && needsSlash) {
     return `${inputPath}/`;
   } else {
@@ -115,3 +117,21 @@ export function getFileWithExtensions(
   }
   return null;
 }
+
+/** Get the Metro server root, when working in monorepos */
+function getMetroServerRoot(projectRoot: string): string {
+  if (env.EXPO_NO_METRO_WORKSPACE_ROOT) {
+    return projectRoot;
+  }
+
+  return findWorkspaceRoot(projectRoot) ?? projectRoot;
+}
+
+/**
+ * Resolve the entry point relative to either the server or project root.
+ * This relative entry path should be used to pass non-absolute paths to Metro,
+ * accounting for possible monorepos and keeping the cache sharable (no absolute paths).
+ */
+export const resolveRelativeEntryPoint: typeof resolveEntryPoint = (projectRoot, options) => {
+  return path.relative(getMetroServerRoot(projectRoot), resolveEntryPoint(projectRoot, options));
+};